pyproject.toml
[project]
name = "quantwinmt5feeder"
version = "0.1.0"
description = "Reads OHLC (M1) bars from MetaTrader 5 and pushes them to the QuantWin REST API."
authors = [
    {name = "Luis", email = "luismaiasombra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "MetaTrader5 (>=5.0.45)",
    "requests (>=2.31.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "tzdata (>=2024.1)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.0,<10.0.0)"
]
